feat: detect foreign key changes and index changes
The  commit contains changes to generate sql scripts for foreign key changes as well as indexes.
We specifically assume db to be postgres i.e by using pg_indexes, querying pg metadata.

A wrapper for `gormDb.Migrator` was added to isolate the usage to make things straightforward to change when gorm directly supports `GetIndexes` and `GetRelationships`